Fix [#284] interceptor 오류 해결
